dev/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.34.39
botocore==1.34.39
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
emoji==2.10.1
exceptiongroup==1.2.0
fastapi==0.109.2
filetype==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langdetect==1.0.9
langsmith==0.0.87
lxml==5.1.0
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pinecone-client==3.0.3
pydantic==2.6.1
pydantic_core==2.16.2
pypdf==4.0.2
python-dateutil==2.8.2
python-iso639==2024.2.7
python-magic==0.4.27
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.36.3
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
unstructured==0.12.4
unstructured-client==0.18.0
urllib3==2.0.7
uvicorn==0.27.0.post1
wrapt==1.16.0
yarl==1.9.4
